sei: move transfer to sei paylaod build logic to the sdk
